gage_study::anova
Struct Anova

pub struct Anova {

    pub n_parts: usize,
    pub n_operators: usize,
    pub n_replicates: usize,
    pub dof_total: usize,
    pub dof_parts: usize,
    pub dof_operators: usize,
    pub dof_repeatability: usize,
    pub dof_part_operator: usize,
    pub sumsq_total: f64,
    pub sumsq_parts: f64,
    pub sumsq_operators: f64,
    pub sumsq_repeatability: f64,
    pub sumsq_part_operator: f64,
    pub meansq_parts: f64,
    pub meansq_operators: f64,
    pub meansq_repeatability: f64,
    pub meansq_part_operator: f64,
    pub f_parts: f64,
    pub f_operators: f64,
    pub f_part_operator: f64,
    pub use_interaction: bool,

}

Constructs all data needed for the ANOVA table

Fields
n_parts: usize

Number of parts: $p$

n_operators: usize

Number of operators: $q$

n_replicates: usize

Number of replicates: $n$

dof_total: usize

Total degrees of freedom: $\nu_T = N - 1$, where $N$ is the number of measurements

dof_parts: usize

Part degrees of freedom: $\nu_p = p - 1$, where $p$ is the number of parts

dof_operators: usize

Operator degrees of freedom: $\nu_q = q - 1$, where $q$ is the number of operators

dof_repeatability: usize

Repeatability degrees of freedom: $\nu_R = pq(n - 1)$, where $n$ is the number of replicates. If the part-operator term is neglected, $\nu_R = \nu_T - \nu_p - \nu_q$.

dof_part_operator: usize

Part-operator interaction degrees of freedom: $\nu_{p \cdot R} = \nu_p * \nu_R$

sumsq_total: f64

Sum of squared differences between every value and the grand mean $$ SS_T = \sum_{i=1}^{p} \sum_{j=1}^{q} \sum_{k=1}^{n} \left(x_{ijk} - \bar{x}_{…} \right)^2 $$

sumsq_parts: f64

Sum of squared differences between each part mean and the grand mean $$ SS_p = qn \sum_{i=1}^{p} \left( \bar{x}_{i..} - \bar{x}_{…} \right)^2 $$

sumsq_operators: f64

Sum of squared differences between each operator mean and the grand mean $$ SS_q = pn \sum_{j=1}^{q} \left( \bar{x}_{.j.} - \bar{x}_{…} \right)^2 $$

sumsq_repeatability: f64

Sum of squared differences between each replicate and the replicate mean $$ SS_R = \sum_{i=1}^{p} \sum_{j=1}^{t} \sum_{k=1}^{q} \left(x_{ijk} - \bar{x}_{ij.} \right)^2 $$ If the part-operator term is neglected, $SS_R = SS_T - SS_p - SS_q$

sumsq_part_operator: f64

Sum of squared differences for the part-operator interaction: $SS_R = SS_T - SS_p - SS_q$

meansq_parts: f64

Mean of the squared differences for parts: $MS_p = \frac{SS_p}{\nu_p}$

meansq_operators: f64

Mean of the squared differences for operators: $MS_q = \frac{SS_q}{\nu_q}$

meansq_repeatability: f64

Mean of the squared differences for repeatability: $MS_R = \frac{SS_R}{\nu_R}$

meansq_part_operator: f64

Mean of the squared differences for the part-operator interaction: $MS_{p \cdot R} = \frac{SS_{p \cdot R}}{\nu_{p \cdot R}}$

f_parts: f64

F-statistic for parts: $F_p = \frac{MS_p}{MS_{p \cdot R}}$. If the part-operator term is negelected $F_p = \frac{MS_p}{MS_R}$.

f_operators: f64

F-statistic for operators: $F_q = \frac{MS_q}{MS_{p \cdot R}}$. If the part-operator term is negelected $F_q = \frac{MS_q}{MS_R}$.

f_part_operator: f64

F-statistic for the part-operator interaction: $F_q = \frac{MS_q}{MS_R}$

use_interaction: bool

Flag for neglecting the part-operator interaction term

Implementations

impl Anova

pub fn new() -> Self

Makes a new AnovaTable

pub fn from_data(dataset: &DataSet) -> Self

Performs all necessary ANOVA calculations using data from a DataSet and populates the AnovaTable fields